created global host launch hook
